refactor: 프로젝트 데이터 호출 관련 로직 훅으로 분리
